requirements.txt
gear>=0.5.4,<1.0.0
python-swiftclient
python-keystoneclient
argparse

lockfile
python-daemon<2.0
extras
GitPython>=0.3.2.RC1
sphinxcontrib-programoutput
sphinxcontrib-seqdiag

mysql-python

requests
PyYAML>=3.1.0,<4.0.0

jenkins-job-builder
xmltodict
python-magic
